move itinerary pdf function to helper file
